+++ src/abi.rs
@@ -1,4 +1,5 @@
 use std::fmt;
+use wasm_contract::Contract;
 
 /// The ABI is a hint to WebAssembly runtimes about what additional imports to insert.
 /// It currently is only used for validation (in the validation subcommand).  The default value is `None`.
@@ -29,5 +30,15 @@ impl fmt::Display for Abi {
 impl Default for Abi {
     fn default() -> Self {
         Abi::None
+    }
+}
+
+impl Abi {
+    pub fn get_contract(&self) -> Option<Contract> {
+        match self {
+            Abi::Emscripten => None,
+            Abi::Wasi => None,
+            Abi::None => None,
+        }
     }
 }