+++ src/sql/queries/wasm_contract_existence_check.sql
@@ -1,0 +1,5 @@
+SELECT 1
+FROM wasm_contracts
+WHERE contract_name = (?1) 
+   AND version = (?2)
+LIMIT 1